requirements.txt
requests~=2.31.0
Pillow~=10.1.0
bs4~=0.0.1
beautifulsoup4~=4.12.2
selenium~=4.16.0
fpdf~=1.7.2
urllib3~=2.1.0
keyboard